requirements.txt
black
python-dotenv
requests
